[feat] Defocus Blur

+++ src/scene_info.cpp
@@ -1,4 +1,4 @@
 ﻿#include "scene_info.h"
 
-int scene_info::vertices = 0;
-int scene_info::faces = 0;
+size_t scene_info::vertices = 0;
+size_t scene_info::faces = 0;
